backend/requirements.txt
fastapi~=0.115
uvicorn[standard]~=0.30
python-multipart~=0.0
pydantic>=2,<3
passlib~=1.7
pyjwt~=2.9
google-cloud-firestore~=2.16
google-auth~=2.35
redis~=5.0
minio~=7.2
prometheus-client~=0.20
psutil
